requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
google-generativeai
duckduckgo-search
pydantic
